:recycle: 컨텐츠 카드 위젯에서 컨텐츠의 개념을 분리한다
